problem 6 cast to int

+++ 1-10.py
@@ -97,8 +97,8 @@ def p6():
   nums = range(1, count + 1)
   sum_of_squares = 0
   for i in nums:
-    sum_of_squares += math.pow(i, 2)
-  square_of_sum = math.pow(sum(nums), 2)
+    sum_of_squares += int(math.pow(i, 2))
+  square_of_sum = int(math.pow(sum(nums), 2))
   print square_of_sum - sum_of_squares
 # or get the formula f(n) = 1^2 + ... + n^2
